Minor adjustments to _getAudioTracksWithHighestEfficiency logic
